Fix broken tests cases
In Correct_Password_Wrong_Encoding_Should_Not_Match a different encoding
is passed to GetHash and CheckPassword. Unfortunately HMAC-SHA1 and
HMAC-256 do not use encoding so the tests fail.
This commit adds the SkippableFact package which allow to skip test
cases base on given conditions and those test cases are skipped.